requirements.txt
# --- Core AI & SetFit (Updated for Compatibility) ---
setfit==1.1.3
transformers==4.41.0
torch==2.2.1
sentence-transformers==2.7.0

# --- UI & Data ---
gradio==4.21.0
pandas==2.2.1
rapidfuzz==3.0.0

# --- Vision & OCR ---
easyocr==1.7.1
pillow==10.2.0

# --- Utils ---
multiprocess==0.70.16

indic-transliteration